Fixed compilation warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@103431 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netcache/active_handler.cpp
@@ -2366,7 +2366,7 @@ CNCActiveHandler::CheckCommandTimeout(void)
         return;
     }
 
-    if (CSrvTime::CurSecs() - m_LastActive
+    if (Uint8(CSrvTime::CurSecs() - m_LastActive)
         > (CNCDistributionConf::GetNetworkErrorTimeout() / kUSecsPerSecond)) {
 #ifdef _DEBUG
 CNCAlerts::Register(CNCAlerts::eDebugSyncAborted2, "CheckCommandTimeout");
